the rate is not calculated since the beggining anymore and a playerToRate variable has been introduced
Calculating the rate everytime a game is created is probably slowing down a lot the program and is only leaded in the end nodes, therefore from now it will be called in the algorithm.
A new variable is introduced so that we always calculate the rate of the same player.