Code reorganization, rating changed and multiple searchs
The code that was not useful anymore has been deleted and the rating algorithm has been modify in order to give less points if the group is of a size different of 3 proportionaly.

The code to create multiples searches in a row to calculate the avg time is introduced in this version.